mywash_admin/requirements.txt
Flask
Werkzeug
Flask-API
virtualenv
flask-restful
pymongo==2.7.2
requests>=2.4
flask-script
Flask-WTF
sendgrid==1.4.0
flask-sqlalchemy==2.0
flask-migrate==1.4.0
googlemaps==2.2
google-api-python-client==1.4.0
psycopg2==2.6
redis==2.10.3
celery==3.1.18
celery-with-redis==3.0
kombu==3.0.24
Jinja2==2.7.3
boto==2.38.0
WeasyPrint==0.23
pycrypto==2.6.1
Flask-Cors==2.0.0
requests-futures==0.9.5
raven==5.7.2
blinker==1.2